Ticket List/Scene Improvements
Created a method in the communicate class that pulls all active tickets for a given tech user name and parses the data into a ListArray of Ticket objects.  Added an update to all scenes that include a "Open Ticket(s)" field to update automatically upon login.  Create global variables for  the ticketList, activeTicketCount, and unreadMessageCount so they could be referenced from any scene/controller/method as needed.