PHP/communicate/currentNote
Update PHP code to return simplified data structure for returning the latest note on a ticket.  Added Note class to hold note data. Updated communicate class to include a method to use the PHP file and parse the response to create a Note object.  Added method to initialization of ViewTicketCommentController to generate a Note object of the latest note of a given ticket.