List of Admins/Assign Ticket Buttons
Added Assign Ticket Form that provides a list of admin users that can have the ticket assigned to them.  Added buttons to use the form and buttons to do a simple unassign/assign to me.  Added communicate method to return an ArrayList of users to hold the admins and created a php script to handle the request.