Command Center Timer/Widget Technician Assignment Update
The Command Center now has a fully implemented timer that updates all scene items without reloading the entire scene.  A global variable was created in Main for the amount of time between updates.  The Ticket Widget viewTickets scene will now list the username of the technician assigned to each ticket.  If unassigned, "Unassigned" will be listed as the username.