Close ticket with reply/(Re)Assign Note/Open Ticket User Selection
When you close a ticket, you can now type a note prior to clicking "Close" and it will add a note prior to closing the ticket.  You can now add a note to a ticket when you assign or reassign the ticket.  When opening a ticket there is now a list of all users you can open a ticket for.  By default, the currently logged in user is selected.